requirements.txt
alembic==1.11.0
aiosqlite==0.19.0
discord.py==2.2.3
emoji==2.2.0
mcrcon==0.7.0
python-dotenv==1.0.0
SQLAlchemy[asyncio]==2.0.13
